Implement seeding the database with EFCore.BulkExtensions
In my test, seeding with EFCore.BulkExtensions is ~2×  slower than fastest method (SeedPicturesWithCommand) which is still waaaaay faster than the standard EFCore AddRange method.

On the plus side, it's literally a one-liner on the DbContext.

Under the hood, it dynamically generates SQL and executes it, just like SeedPicturesWithCommand but with a little bit more overhead.